Flytt forretningslogikk inn i en service
