+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/familieintegrasjoner/IntegrasjonClient.kt
@@ -1,6 +1,5 @@
 package no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner
 
-import no.nav.familie.ba.sak.common.MDCOperations
 import no.nav.familie.ba.sak.common.kallEksternTjenesteRessurs
 import no.nav.familie.ba.sak.common.kallEksternTjenesteUtenRespons
 import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.domene.Arbeidsfordelingsenhet
@@ -15,18 +14,12 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.task.OpprettTaskService.Companion.RETRY_BACKOFF_5000MS
 import no.nav.familie.http.client.AbstractRestClient
-import no.nav.familie.http.client.RessursException
-import no.nav.familie.kontrakter.felles.BrukerIdType
 import no.nav.familie.kontrakter.felles.Fagsystem
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.dokarkiv.ArkiverDokumentResponse
 import no.nav.familie.kontrakter.felles.dokarkiv.Dokumenttype
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.ArkiverDokumentRequest
-import no.nav.familie.kontrakter.felles.dokarkiv.v2.Dokument
-import no.nav.familie.kontrakter.felles.dokarkiv.v2.Filtype
-import no.nav.familie.kontrakter.felles.dokarkiv.v2.Førsteside
 import no.nav.familie.kontrakter.felles.dokdist.DistribuerJournalpostRequest
-import no.nav.familie.kontrakter.felles.journalpost.Bruker
 import no.nav.familie.kontrakter.felles.journalpost.Journalpost
 import no.nav.familie.kontrakter.felles.journalpost.JournalposterForBrukerRequest
 import no.nav.familie.kontrakter.felles.kodeverk.KodeverkDto
@@ -41,7 +34,6 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpStatus
 import org.springframework.retry.annotation.Backoff
 import org.springframework.retry.annotation.Retryable
 import org.springframework.stereotype.Component
@@ -351,98 +343,19 @@ class IntegrasjonClient(
         }
     }
 
-    fun journalførManueltBrev(
-        fnr: String,
-        fagsakId: String,
-        journalførendeEnhet: String,
-        brev: ByteArray,
-        dokumenttype: Dokumenttype,
-        førsteside: Førsteside?
-    ): String {
-        return journalførDokument(
-            fnr = fnr,
-            fagsakId = fagsakId,
-            journalførendeEnhet = journalførendeEnhet,
-            brev = listOf(
-                Dokument(
-                    dokument = brev,
-                    filtype = Filtype.PDFA,
-                    dokumenttype = dokumenttype
-                )
-            ),
-            førsteside = førsteside,
-        )
-    }
-
     fun journalførDokument(
-        fnr: String,
-        fagsakId: String,
-        journalførendeEnhet: String? = null,
-        brev: List<Dokument>,
-        vedlegg: List<Dokument> = emptyList(),
-        førsteside: Førsteside? = null,
-        behandlingId: Long? = null
-    ): String {
+        arkiverDokumentRequest: ArkiverDokumentRequest
+    ): ArkiverDokumentResponse {
         val uri = URI.create("$integrasjonUri/arkiv/v4")
-        logger.info("Sender pdf til DokArkiv: $uri")
 
-        if (journalførendeEnhet == DEFAULT_JOURNALFØRENDE_ENHET) {
-            logger.warn("Informasjon om enhet mangler på bruker og er satt til fallback-verdi, $DEFAULT_JOURNALFØRENDE_ENHET")
+        return kallEksternTjenesteRessurs(
+            tjeneste = "dokarkiv",
+            uri = uri,
+            formål = "Journalfør dokument på fagsak ${arkiverDokumentRequest.fagsakId}"
+        ) {
+            postForEntity(uri, arkiverDokumentRequest)
         }
-
-        val eksternReferanseId = "${fagsakId}_${behandlingId}_${MDCOperations.getCallId()}"
-
-        val journalpostId = try {
-            val journalpost = kallEksternTjenesteRessurs(
-                tjeneste = "dokarkiv",
-                uri = uri,
-                formål = "Journalfør dokument på fagsak $fagsakId",
-            ) {
-                val arkiverDokumentRequest = ArkiverDokumentRequest(
-                    fnr = fnr,
-                    forsøkFerdigstill = true,
-                    hoveddokumentvarianter = brev,
-                    vedleggsdokumenter = vedlegg,
-                    fagsakId = fagsakId,
-                    journalførendeEnhet = journalførendeEnhet,
-                    førsteside = førsteside,
-                    eksternReferanseId = eksternReferanseId
-                )
-
-                postForEntity<Ressurs<ArkiverDokumentResponse>>(uri, arkiverDokumentRequest)
-            }
-
-            if (!journalpost.ferdigstilt) {
-                error("Klarte ikke ferdigstille journalpost med id ${journalpost.journalpostId}")
-            }
-
-            journalpost.journalpostId
-        } catch (ressursException: RessursException) {
-            when (ressursException.httpStatus) {
-                HttpStatus.CONFLICT -> {
-                    logger.warn(
-                        "Klarte ikke journalføre dokument på fagsak=$fagsakId fordi det allerede finnes en journalpost " +
-                            "med eksternReferanseId=$eksternReferanseId."
-                    )
-
-                    hentEksisterendeJournalpost(eksternReferanseId, fnr)
-                }
-                else -> throw ressursException
-            }
-        }
-
-        return journalpostId
     }
-
-    private fun hentEksisterendeJournalpost(
-        eksternReferanseId: String,
-        fnr: String
-    ): String = hentJournalposterForBruker(
-        JournalposterForBrukerRequest(
-            brukerId = Bruker(id = fnr, type = BrukerIdType.FNR),
-            antall = 50,
-        )
-    ).single { it.eksternReferanseId == eksternReferanseId }.journalpostId
 
     fun opprettSkyggesak(aktør: Aktør, fagsakId: Long) {
         val uri = URI.create("$integrasjonUri/skyggesak/v1")
